+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Semantic versioning in our case means:
   change the client facing API, change code conventions significantly, etc.
 
 
-## WIP
+## 1.8.1
 
 ### Features
 
@@ -26,7 +26,7 @@ Semantic versioning in our case means:
 
 ### Bugfixes
 
-- Count overused str and bytes separately for `WPS226`, 3782
+- Count overused str and bytes separately for `WPS226`, #3782
 - Do not count docstrings for `WPS226`, #3803
 - Do not count attribute and type alias docstrings for `WPS226`, #3805
 - Adds direct links to violations when searching by code, #3790